# First pass at slightly smarter (and much faster) shuffling - RFC