--- a/src/model/ComputationChild.js
+++ b/src/model/ComputationChild.js
@@ -30,6 +30,11 @@ export default class ComputationChild extends Model {
 		return this.childByKey[ key ];
 	}
 
+	// TODO: this should probably just work off of the parent
+	tryRebind () {
+		return;
+	}
+
 	// TODO this causes problems with inter-component mappings
 	// set () {
 	// 	throw new Error( `Cannot set read-only property of computed value (${this.getKeypath()})` );