--- a/src/model/RootModel.js
+++ b/src/model/RootModel.js
@@ -179,6 +179,11 @@ export default class RootModel extends Model {
 		super.teardown();
 	}
 
+	// root models don't shuffle
+	tryRebind () {
+		return this;
+	}
+
 	update () {
 		// noop
 	}