--- a/src/model/specials/KeypathModel.js
+++ b/src/model/specials/KeypathModel.js
@@ -45,6 +45,12 @@ export default class KeypathModel {
 		this.children.forEach( teardown );
 	}
 
+	tryRebind () {
+		const next = this.parent.tryRebind();
+		if ( next ) return next.getKeypathModel( this.ractive );
+		return next;
+	}
+
 	unregister ( dependant ) {
 		removeFromArray( this.dependants, dependant );
 	}