--- a/src/model/specials/RactiveModel.js
+++ b/src/model/specials/RactiveModel.js
@@ -14,4 +14,9 @@ export default class RactiveModel extends Model {
 	getKeypath() {
 		return '@this';
 	}
+
+	// root models don't shuffle
+	tryRebind () {
+		return false;
+	}
 }