--- a/src/view/items/shared/Mustache.js
+++ b/src/view/items/shared/Mustache.js
@@ -1,5 +1,6 @@
 import Item from './Item';
 import resolve from '../../resolvers/resolve';
+import runloop from '../../../global/runloop';
 
 export default class Mustache extends Item {
 	constructor ( options ) {
@@ -45,9 +46,14 @@ export default class Mustache extends Item {
 	}
 
 	rebind () {
-		if ( this.isStatic ) return;
+		const force = runloop.isForceRebinding();
+		if ( this.isStatic || ( !force && !this.model ) ) return;
 
-		const model = resolve( this.parentFragment, this.template );
+		let model;
+		if ( this.model ) model = this.model.tryRebind();
+
+		if ( model === false ) return;
+		else if ( !model ) model = resolve( this.parentFragment, this.template );
 
 		if ( model === this.model ) return;