--- a/src/view/resolvers/ReferenceExpressionProxy.js
+++ b/src/view/resolvers/ReferenceExpressionProxy.js
@@ -70,10 +70,6 @@ export default class ReferenceExpressionProxy extends Model {
 			this.resolvers.push( baseResolver );
 		}
 
-		const intermediary = {
-			handleChange: () => this.handleChange()
-		};
-
 		this.members = template.m.map( ( template, i ) => {
 			if ( typeof template === 'string' ) {
 				return { get: () => template };
@@ -86,12 +82,12 @@ export default class ReferenceExpressionProxy extends Model {
 				model = resolveReference( fragment, template.n );
 
 				if ( model ) {
-					model.register( intermediary );
+					model.register( this );
 				} else {
 					resolver = fragment.resolve( template.n, model => {
 						this.members[i] = model;
 
-						model.register( intermediary );
+						model.register( this );
 						this.handleChange();
 
 						removeFromArray( this.resolvers, resolver );
@@ -104,7 +100,7 @@ export default class ReferenceExpressionProxy extends Model {
 			}
 
 			model = new ExpressionProxy( fragment, template );
-			model.register( intermediary );
+			model.register( this );
 			return model;
 		});
 
@@ -220,7 +216,33 @@ export default class ReferenceExpressionProxy extends Model {
 
 	// TODO: this should really shuffle
 	tryRebind () {
-		return;
+		let dirty = false;
+
+		if ( this.base ) {
+			const next = this.base.tryRebind();
+			if ( next ) {
+				this.base = next;
+			} else if ( next !== false ) return;
+		}
+
+		this.members.forEach( ( m, i ) => {
+			if ( m ) {
+				const next = m.tryRebind();
+				if ( next ) {
+					this.members.splice( i, 1, next );
+					m.unregister( this );
+					next.register( this );
+				}
+				else if ( next !== false ) dirty = true;
+			}
+		});
+
+		// if something couldn't rebind, tell the caller to re-resolve
+		if ( dirty ) return;
+
+		this.bubble();
+
+		return this;
 	}
 
 	unbind () {