**Description of the pull request:**
Shuffling is one of the last things that is somewhat slow in edge, primarily because when an array shuffles, everything downstream from that array re-resolves all of its references, which is a pretty slow process. This attempts to make that process slightly smarter by assuming that there are roughly three states for a shuffling rebind: not affected by the shuffle, shuffled, and unknown. The unknown state is kind of a special case that is also used by unmapping and remapping in components to get updates propagated correctly.

When a shuffle kicks off, it registers with the runloop. Then when a rebind fires, if it already has a model, it checks to with the model to see if there is a known new model. If there is, it uses it. If it is known that the model is not affected by a shuffle, then it stops there. If it is unknown, then the usual re-resolve process takes place.

Since keypaths are used a ton in this, it also adds keypath caching to avoid a pretty good bit of keypath re-calc. Substitution is still done if the path is mapped, but it doesn't have to start by walking up to the root model.

I have all of the tests passing, but I'm not positive that the test suite covers 100% of the corner cases involved in shuffling. I had to add actual rebind handlers in a few places because there were/are a few items that just unbind/bind on rebind.

There are still a few models and model-like things that do not participate in the tryRebind process, but there shouldn't be any reason why everything wouldn't be covered. Notably, I haven't looked at reference expressions and children of those and computations at all yet. At this point, I think everything in the jsweb perf suite is covered, and I'm seeing this take roughly 35% of the time that edge does for removing the first row (out of 1000) and 20% of the time for the last. Splicing the last row is still not quite as fast as 0.7.3, but everything else is considerably faster.

**Fixes**
#2366 - to the extent that it can... edge is already considerably faster than 0.7.3 in most of the perf suite.

**Is breaking:**
Doesn't look like it, but the test suite may be lacking.

**Reviewers:**
Yes please! If you have a little time to look over the implementation or try out this PR, by all means, please do, and feel free to leave some feedback!