@JonDum 2016-06-21
Very sexy. I'm gonna try this out with `comp-bind` in the morning 😈