@sabob 2016-06-21
Haven't tried this out yet, but over the weekend I noticed the perf test in #2366 was a bit slower that it was couple of months ago:

"create 10000 rows" went up from 500ms to 700ms.
"Delete row" went up from 180ms to 350ms.

Will try out this PR and see the difference.