@evs-chris 2016-06-21
I've played around a bit more with this and learned a few interesting things:

1. A Chromebook Flip is roughly as fast as a Nexus 6P (not very fast)
2. ... which are only about 50% slower than Edge on a kickass Windows development machine at DOM manipulation 😕. For reference, by kickass, I mean runs visual studio 2015 without stuttering _anywhere_ - I know, right?
3. My linux dev box appears to be 20-50% slower than the machine doing the benchmarks mentioned in #2366, based on their current table and depending on the test step.
4. This PR (and edge in general) beats the pants off of 0.7.3 in just about every test (not splice last row).
5. Even just keypath caching seems to contribute a fair amount of performance bump for lots of vdom.
6. A ton of DOM is painful to deal with no matter where you are. Pagination is your friend.
7. The Flip actually gets a kill script popup during the test with 0.7.3.

Here are the relevant numbers on the kickass machine:

|test|edge|pr|0.7.3|
|---:|---:|---:|---:|
|create 1000|177.2|166.7|301.9|
|remove first|66.0|23.9|104.8|
|remove last|47.8|13.1|5.6|
|create 10,000|1846.2|1147.1|3660.9|

... and the linux dev machine:

|test|edge|pr|0.7.3|
|---:|---:|---:|---:|
|create 1000|443.8|393.4|882.6|
|remove first|146.0|73.0|229.7|
|remove last|98.6|37.7|11.8|
|create 10,000|2601.9|2535.3|14112.6|

... and the Chromebook Flip:

|test|edge|pr|0.7.3|
|---:|---:|---:|---:|
|create 1000|1892.8|1588.9|1813.3|
|remove first|315.8|171.8|676.1|
|remove last|214.6|77.8|29.5|
|create 10,000|6872.9|6411.6|20807.1|

All tests were done on Chrome 51, except the Flip, which is on 53 (not that it matters).