@evs-chris 2016-06-21
I should probably have noted that I updated the tests so that the removals are only working with 1000 rows instead of starting with whatever was left after running however many 'add 100 rows' iterations could finish. My original 35% and 20% figures were based on working with ~1700 rows, so this seems to scale a lot better than edge.

Also, the numbers are avg ms from the output of the jsweb perf test.