@JonDum 2016-06-21
Those numbers are insane improvements!

Strengthens my resolve through the dark times of the almost year-long hodgepodge ordeal that the 0.8 refactor has been.