@mefjuu 2016-06-28
@aharpervc Yes, RactiveJS release cycle is the only thing confusing me when I'm picking the library for a project...