@evs-chris 2016-06-28
The good news is that Ractive shouldn't have any more massive refractors like 0.8 so maintenance and further development will be much easier. Also, 0.8 is about to go into alpha, so it should be out soon™.