@aharpervc 2016-06-28
If the 0.8 api is now stable, perhaps it is appropriate to call it 1.0