@skeptic35 2016-07-11
Don't know if I'm right, but I suspect this change to be responsible for breaking ... well, basically just about everything in my application. I was running an 0.8 edge version, built on 13th of June 2016. Everything was looking good, ran beautifully fast, and I was really looking forward to the final 0.8 release. Then today I decided to pull the latest edge version, started up my app and almost nothing worked anymore. And I don't even get any errors. For some reason some of the ```complete``` events don't get fired anymore and some of my observers don't get called anymore. If at least I got some errors, but no, not a single one. 
What led me to suspect that this particular change is responsible, is the fact that at some point I actually did receive an error: "TypeError: m.tryRebind is not a function(…)" in line 8855 of ractive.js
It was the ```tryRebind``` part that led me here. 
I'm really sorry, that I can't provide any helpful input to solve the problem or even narrow it down. As I said: that was the only error I got, possibly initiated by a ```.splice()``` call, some 50 or 60 steps down into the trace log.

It's terribly frustrating! Finally 0.8. seemed to be working just fine and to be getting close to release and now this. What I really don't get is the following: Why introduce this change into 0.8 at all? It doesn't seem related to any bug and the performance of edge already was very much superior to 0.7. So why not just fix the open issues and get 0.8 on the road? It's been long enough already since 0.7.3. No need to slow down the process by introducing massive changes like this one, which always bear the risk to break things.

Sorry for that rant, I'm usually a big fan of Ractive and I really appreciate all the great work you guys do but seeing my app fall apart with the latest edge version was simply the single most frustrating thing that
Ractive ever did to me.