@michaljerabek 2016-07-11
The bug with `oncomplete` is caused by https://github.com/ractivejs/ractive/issues/2609.