@evs-chris 2016-07-11
I'm reverting this to the old rebinding behavior. I'm not positive that this actually caused the issues or if another bug fix in the interim did, but either way, I've been looking at a number of other somewhat related issues and have come to the conclusion that rebinding needs to happen from the model side, as it appears to in #1740, rather than from the view, as it does now.

Performance may be better with edge than with 0.7, but splicing around lists of any significant length is way slower than it should be. I consider it a bug when the page freezes for a second or two when removing an item from the middle of a list of 1500 items.

_I'll_ consider edge frozen, except bug fixes only now, and everything else will wait until 0.9, unless another maintainer or BDFL wants it in.