@skeptic35 2016-07-19
Obviously the fact that my complete handlers were not called was caused by commit 8a17551. But the TypeError for```m.tryRebind``` remains, even if I use a pached version of Ractive including this pull request.
I do think that it is the right decision to postpone this till after 0.8. I'll be glad to help debugging it then, if I can. After all it's not like I'm not in favor of performance improvements. And the last thing I want is to cause frustration on your side :)